requirements.txt
brotlipy==0.7.0
certifi==2022.5.18.1
Flask==2.1.2
Flask-Cors==3.0.10
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.1.0
MarkupSafe==2.1.1
munkres==1.1.4
numpy==1.22.3
python-dotenv==0.20.0
requests==2.28.0
six==1.16.0
unicodecsv==0.14.1
urllib3==1.26.7
webencodings==0.5.1
Werkzeug==2.1.2